Rename TestNetworkTransportDispatcher -> TestNetworkTransportHandler
