+++ tests/test-network-transport/src/test/kotlin/testnetworktransport/QueueTestNetworkTransportHandlerTest.kt
@@ -4,9 +4,9 @@ import com.apollographql.apollo3.ApolloClient
 import com.apollographql.apollo3.annotations.ApolloExperimental
 import com.apollographql.apollo3.api.ApolloResponse
 import com.apollographql.apollo3.api.Error
+import com.apollographql.apollo3.testing.QueueTestNetworkTransportHandler
 import com.apollographql.apollo3.testing.TestNetworkTransport
 import com.apollographql.apollo3.testing.runTest
-import com.apollographql.apollo3.testing.testNetworkTransport
 import com.benasher44.uuid.uuid4
 import testnetworktransport.test.GetHeroQuery_TestBuilder.Data
 import kotlin.test.Test
@@ -15,12 +15,14 @@ import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 @OptIn(ApolloExperimental::class)
-class MapTestNetworkTransportDispatcherTest {
+class QueueTestNetworkTransportHandlerTest {
+  private lateinit var queue: QueueTestNetworkTransportHandler
   private lateinit var apolloClient: ApolloClient
 
   private fun setUp() {
+    queue = QueueTestNetworkTransportHandler()
     apolloClient = ApolloClient.Builder()
-        .networkTransport(TestNetworkTransport())
+        .networkTransport(TestNetworkTransport(queue))
         .build()
   }
 
@@ -29,7 +31,7 @@ class MapTestNetworkTransportDispatcherTest {
   }
 
   @Test
-  fun registerResponses() = runTest(before = { setUp() }, after = { tearDown() }) {
+  fun enqueueResponses() = runTest(before = { setUp() }, after = { tearDown() }) {
     val query1 = GetHeroQuery("001")
     val testResponse1 = ApolloResponse.Builder(query1, uuid4(), null)
         .errors(listOf(Error(
@@ -56,10 +58,9 @@ class MapTestNetworkTransportDispatcherTest {
     val testResponse3 = ApolloResponse.Builder(query3, uuid4(), GetHeroNameOnlyQuery.Data(GetHeroNameOnlyQuery.Hero(name = "Darth Vader")))
         .build()
 
-    val testNetworkTransport = apolloClient.testNetworkTransport
-    testNetworkTransport.register(query1, testResponse1)
-    testNetworkTransport.register(query2, testResponse2)
-    testNetworkTransport.register(query3, testResponse3)
+    queue.enqueue(testResponse1)
+    queue.enqueue(testResponse2)
+    queue.enqueue(testResponse3)
 
     val actual1: ApolloResponse<GetHeroQuery.Data> = apolloClient.query(query1).execute()
     assertTrue(actual1.hasErrors())
@@ -75,9 +76,9 @@ class MapTestNetworkTransportDispatcherTest {
   }
 
   @Test
-  fun registerError() = runTest(before = { setUp() }, after = { tearDown() }) {
+  fun enqueueError() = runTest(before = { setUp() }, after = { tearDown() }) {
     val query = GetHeroQuery("001")
-    apolloClient.testNetworkTransport.register(query, errors = listOf(Error(
+    queue.enqueue(query, errors = listOf(Error(
         message = "There was an error",
         locations = listOf(Error.Location(line = 1, column = 2)),
         path = listOf("hero", "name"),
@@ -92,7 +93,7 @@ class MapTestNetworkTransportDispatcherTest {
   }
 
   @Test
-  fun registerDataManual() = runTest(before = { setUp() }, after = { tearDown() }) {
+  fun enqueueDataManual() = runTest(before = { setUp() }, after = { tearDown() }) {
     val query = GetHeroQuery("001")
     val testData = GetHeroQuery.Data(
         GetHeroQuery.Hero(
@@ -103,21 +104,21 @@ class MapTestNetworkTransportDispatcherTest {
             onHuman = null
         )
     )
-    apolloClient.testNetworkTransport.register(query, testData)
+    queue.enqueue(query, testData)
 
     val actual = apolloClient.query(query).execute().data
     assertEquals(testData, actual)
   }
 
   @Test
-  fun registerDataTestBuilder() = runTest(before = { setUp() }, after = { tearDown() }) {
+  fun enqueueDataTestBuilder() = runTest(before = { setUp() }, after = { tearDown() }) {
     val query = GetHeroQuery("001")
     val testData = GetHeroQuery.Data {
       hero = droidHero {
         name = "R2D2"
       }
     }
-    apolloClient.testNetworkTransport.register(query, testData)
+    queue.enqueue(query, testData)
 
     val actual = apolloClient.query(query).execute().data!!
     assertEquals(testData.hero.name, actual.hero.name)
